+++ lib/run/run.command.js
@@ -2,9 +2,9 @@ const path = require('path')
 
 const { log } = require('../shared/log.utils')
 
-const { ConfigurationFileNotExist } = require('../services/config-file-not-exist.error')
-const { ScriptNotExist } = require('./script-not-exist.error')
-const { BadScriptPermission } = require('./bad-script-permission.error')
+const { ConfigurationFileNotExist } = require('../services/config/config-file-not-exist.error')
+const { ScriptNotExist } = require('../services/file/script-not-exist.error')
+const { BadScriptPermission } = require('../services/file/bad-script-permission.error')
 
 module.exports = exports = (configService, fileService) => (scriptName, options) => {
   Promise.resolve()
