+++ lib/services/config/config.service.js
@@ -1,7 +1,7 @@
 const fs = require('fs')
 
 const { ConfigurationFileNotExist } = require('./config-file-not-exist.error')
-const { FsUtils } = require('../shared/fs.utils')
+const { FsUtils } = require('../../shared/fs.utils')
 
 class ConfigService {
   get repoUrl() {
