+++ lib/services/file/file.service.js
@@ -1,12 +1,12 @@
 const fs = require('fs')
 const path = require('path')
 
-const { FsUtils } = require('../shared/fs.utils')
-const { ProcessUtils } = require('../shared/process.utils')
-const { log } = require('../shared/log.utils')
+const { FsUtils } = require('../../shared/fs.utils')
+const { ProcessUtils } = require('../../shared/process.utils')
+const { log } = require('../../shared/log.utils')
 
-const { ScriptNotExist } = require('../run/script-not-exist.error')
-const { BadScriptPermission } = require('../run/bad-script-permission.error')
+const { ScriptNotExist } = require('./script-not-exist.error')
+const { BadScriptPermission } = require('./bad-script-permission.error')
 
 class FileService {
   get scripts() {
