+++ lib/services/file/script-not-exist.error.js
